Use restic from alpine
Backrest retrieves the restic package if no environment variable to the binary is set and places this in /data which is a docker volume. In my case but perhaps not limited to me this is a noexec mount  and/or it can clash with selinux. 

This patch  installs  restic from alpine in the docker image.