feat!: redefine hostname as a required property that maps to --host
This is a breaking change with existing backrest behavior. On next
access of the backrest UI clients will be required to set an explicit
hostname for each backrest node. Host names are immutable (only
changeable by manually editing the config) and are associated with all
snapshots.

Backrest will only manage snapshots with a hostname tag matching the
current host. A migration plan for existing snapshots will be included
in a followup.