Clarify notes about the port for Windows installation
Apologies if I caused confusion around the port behaviour. Running the installer as you described with PowerShell would have the installer pick up the variable and use the port for UI shortcuts, but Backrest would still listen on the default port.
Also, using :<port> instead of 127.0.0.1:<port> would cause Backrest to listen on all interfaces, thus exposing it to outside access.
The correct way to define a user variable persistently is described in this commit.
I also made changes to the installer, including the language around the port.
If Backrest supported port definition through a command-line switch (like --port=<port>), then I could add an option to the installer to take user input and create the startup shortcut with the switch. The switch would have to be available for the backrest-windows-tray.exe since that's what starts the software.
Not saying this is needed, just explaining what would have to be done to avoid the environment variable steps I described. Unfortunately, Windows doesn't support passing session variables  through shortcuts, unless using cmd wrapper script.