Mitigates loss of snapshots by requiring --keep-last n restic param for sub-hourly cron schedules
Adds a new field to control the --keep-last parameter and offers a calculation of min and max duration for n snapshots in the current schedule. This provides a way to keep backups more frequent than hourly and offers a cue to the user that the retention settings are not sufficient for their cron.