fix: overwrite service scripts
Both the systemd service unit and the openrc
service script are now overwritten in any case.
Two different bugs arise from the previous version:
1. The service scripts couldn't be updated
automatically along with the rest of the software.
This includes changing 0.0.0.0 with 127.0.0.1.
2. The functions that copy the new scripts and
then restart the service were not executed,
so the service was not re-started after the update.